+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## ???
+- Support LaTeX compiling using the LaTeX Workshop extension
 - Log to output channel
 - Fix synchronization bug (from project copy to working directory)